requirements-pi.txt
pyzmq
numpy
python-dotenv
